@@ -12,32 +12,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.0",
-    "@angular/cli": "^7.2.0",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/forms": "^7.2.0",
-    "@angular/http": "^7.2.0",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular/platform-browser-dynamic": "^7.2.0",
-    "@angular/router": "^7.2.0",
+    "@angular/animations": "^8.2.4",
+    "@angular/cli": "^8.3.2",
+    "@angular/common": "^8.2.4",
+    "@angular/compiler": "^8.2.4",
+    "@angular/core": "^8.2.4",
+    "@angular/forms": "^8.2.4",
+    "@angular/platform-browser": "^8.2.4",
+    "@angular/platform-browser-dynamic": "^8.2.4",
+    "@angular/router": "^8.2.4",
     "big.js": "^5.0.2",
     "bip38": "^2.0.2",
     "bip39": "^2.5.0",
     "bitcoinjs-lib": "^5.1.6",
+    "buffer": "^5.4.2",
     "core-js": "^2.5.4",
     "material-design-icons": "^3.0.1",
     "materialize-css": "^1.0.0",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.3",
     "tslib": "^1.9.0",
     "wif": "^3.0.1",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/compiler-cli": "^7.2.0",
-    "@angular/language-service": "^7.2.0",
+    "@angular-devkit/build-angular": "~0.803.2",
+    "@angular/compiler-cli": "^8.2.4",
+    "@angular/language-service": "^8.2.4",
     "@types/big.js": "^4.0.0",
     "@types/bip32": "~1.0.0",
     "@types/bip39": "2.4.0",
@@ -46,7 +46,7 @@
     "@types/materialize-css": "^1.0.1",
     "@types/node": "~8.9.4",
     "@types/wif": "~2.0.1",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
@@ -57,6 +57,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.5.3"
   }
 }
